refactor: update analysis and entity tools to use SQLModel
- Convert headline_analyzer.py to use SQLModel syntax
- Update entity_resolver.py to use SQLModel for entity creation
- Update entity_tracker.py to use SQLModel for all entity operations
- Replace all SQLAlchemy Session imports with SQLModel Session
- Replace Pydantic Create models with SQLModel classes
